Move signed_compare from utility to math
